backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
spotipy==2.23.0
python-dotenv==1.0.1
sqlalchemy==1.4.48
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
unidecode==1.4.0
